Adding payment login test
