+++ src/test/java/TestPackage/navigationPageTest/NavigationToPaymentPage.java
@@ -1,4 +1,4 @@
-package TestPackage.NavigationPage;
+package TestPackage.navigationPageTest;
 
 import org.testng.annotations.Test;
 
@@ -7,10 +7,10 @@ import com.customize.reporter.WebReporter;
 import Page.NavigationPage;
 import TestCases.AbstractTestCases;
 
-public class NavigationToNewListPage extends AbstractTestCases {
+public class NavigationToPaymentPage extends AbstractTestCases {
 
-    @Test(testName = "5", description = "通过导航栏进入到新闻动态页面")
-    public void navigationToNewListPage() throws Exception {
+    @Test(testName = "3", description = "通过导航栏进入到在线充值页面")
+    public void navigationToPaymentPage() throws Exception {
         NavigationPage navigationPage = new NavigationPage(driver);
         navigationPage.clickPayPageLink();
         WebReporter.log(driver, driver.getTitle(), true, true);
